Move Zoom to separate module. Rework Zoom.equals to Zoom.mapKey.
